requirements.txt
pyyaml
boto3
python-twitter
yfinance
lxml
